+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/2.8/applet.js
@@ -379,7 +379,7 @@ SpicesUpdate.prototype = {
       "general_frequency",
       this.on_frequency_changed,
       null);
-      this.refreshInterval = 3600*this.general_frequency;
+      this.refreshInterval = 3600 * this.general_frequency;
 
     this.settings.bindProperty(Settings.BindingDirection.IN,
       "general_warning",
@@ -644,7 +644,7 @@ SpicesUpdate.prototype = {
       Mainloop.source_remove(this.loopId);
     }
     this.loopId = 0;
-    this.refreshInterval = 3600*this.general_frequency;
+    this.refreshInterval = 3600 * this.general_frequency;
     this.loopId = Mainloop.timeout_add_seconds(this.refreshInterval, () => this.updateLoop());
   }, // End of on_frequency_changed
 
@@ -659,7 +659,7 @@ SpicesUpdate.prototype = {
     this._set_main_label();
 
     // Refresh intervall:
-    this.refreshInterval = 3600*this.general_frequency;
+    this.refreshInterval = 3600 * this.general_frequency;
 
     // Types to check
     this.types_to_check = [];
@@ -1356,7 +1356,7 @@ SpicesUpdate.prototype = {
         var nearest_commit_time = timestamp;
         var smaller_difference = Math.round(Date.now() / 1000);
         let difference;
-        while (result = subject_regexp.exec(data.toString())) {
+        while (result == subject_regexp.exec(data.toString())) {
           commit_time = Date.parse(result[1].toString()) / 1000;
           difference = Math.abs(timestamp - commit_time);
           if (difference < smaller_difference) {
@@ -1608,10 +1608,17 @@ SpicesUpdate.prototype = {
       this.spicesMenuItems[t].setShowDot(this.menuDots[t]);
       this.menu.addMenuItem(this.spicesMenuItems[t]);
     }
+    // button Forget
     if (this.nb_to_watch > 0) {
       let _forget_button = new PopupMenu.PopupIconMenuItem(_("Forget new Spices") + " -\u2604-", "emblem-ok", St.IconType.SYMBOLIC);
-      _forget_button.connect("activate", (event) => this._on_forget_new_spices_pressed())
+      _forget_button.connect("activate", (event) => this._on_forget_new_spices_pressed());
       this.menu.addMenuItem(_forget_button);
+    }
+    //button Download
+    if ((this.nb_to_update + this.nb_to_watch) > 0) {
+      let _download_tabs_button = new PopupMenu.PopupIconMenuItem(_("Open useful Cinnamon Settings"), "folder-download", St.IconType.SYMBOLIC);
+      _download_tabs_button.connect("activate", (event) => this.open_each_download_tab());
+      this.menu.addMenuItem(_download_tabs_button);
     }
     this.menu.addMenuItem(new PopupMenu.PopupSeparatorMenuItem());
 
@@ -1623,7 +1630,7 @@ SpicesUpdate.prototype = {
     this.menu.addMenuItem(configure);
     if (DEBUG) {
       let _reload_button = new PopupMenu.PopupIconMenuItem("Reload this applet", "edit-redo", St.IconType.SYMBOLIC);
-      _reload_button.connect("activate", (event) => this._on_reload_this_applet_pressed())
+      _reload_button.connect("activate", (event) => this._on_reload_this_applet_pressed());
       this.menu.addMenuItem(_reload_button);
     }
 
@@ -1685,18 +1692,26 @@ SpicesUpdate.prototype = {
     }
     if (this.nb_to_update > 0 || this.nb_to_watch > 0) {
       var _tooltip = this.default_tooltip;
+      var tooltip_was_modified = false;
       for (let type of TYPES) {
         if (this.old_message[type] != "" || this.old_watch_message[type] != "") {
+          if (!tooltip_was_modified) {
+            _tooltip += "\n%s".format(_("Middle-Click to open useful Cinnamon Settings"));
+            tooltip_was_modified = true;
+          }
           _tooltip += "\n\n\t\t\t%s".format(_(type).toLocaleUpperCase());
           if (this.old_message[type] != "") _tooltip += "\n\u21BB %s".format(this._clean_str(this.old_message[type].replace(/, /gi, "\n\t")));
           if (this.old_watch_message[type] != "") _tooltip += "\n\u2604 %s".format(this._clean_str(this.old_watch_message[type].replace(/, /gi, "\n\t")));
         }
       }
+      if (!tooltip_was_modified) {
+        _tooltip += "\n%s".format(_("Middle-Click to Refresh"));
+      }
       this.set_applet_tooltip(_tooltip);
       this.numberLabel.text = (this.nb_to_update + this.nb_to_watch).toString();
       this.badge.show();
     } else {
-      this.set_applet_tooltip(this.default_tooltip);
+      this.set_applet_tooltip(this.default_tooltip + "\n%s".format(_("Middle-Click to Refresh")));
       this.numberLabel.text = '';
       this.badge.hide();
     }
@@ -1729,7 +1744,7 @@ SpicesUpdate.prototype = {
         this.refreshInterval = 5;
       } else {
         if (!this.first_loop) {
-          this.refreshInterval = 3600*this.general_frequency;
+          this.refreshInterval = 3600 * this.general_frequency;
           var monitor, Id;
           for (let tuple of this.monitors) {
             [monitor, Id] = tuple;
@@ -1817,6 +1832,43 @@ SpicesUpdate.prototype = {
       this.loopId = Mainloop.timeout_add_seconds(this.refreshInterval, () => this.updateLoop());
     }
   }, // End of updateLoop
+
+  open_each_download_tab:function() {
+    for (let t of TYPES) {
+      if (this.nb_in_menu[t] > 0) {
+        Util.spawnCommandLine("cinnamon-settings %s".format(t.toString()));
+      }
+    }
+  },  // End of open_each_download_tab
+
+  _onButtonPressEvent:function(actor, event) {
+    if (this._applet_enabled) {
+			if (event.get_button() == 1) {
+				if (!this._draggable.inhibit) {
+					return false;
+				} else {
+					if (this._applet_context_menu.isOpen) {
+						this._applet_context_menu.toggle();
+					}
+					this.on_applet_clicked(event);
+				}
+			}
+			if (event.get_button() == 2) {
+	      if ((this.nb_to_update + this.nb_to_watch) === 0) {
+	        this._on_refresh_pressed();
+	      } else {
+	        this.open_each_download_tab();
+	      }
+	    }
+    
+			if (event.get_button() == 3) {
+				if (this._applet_context_menu._getMenuItems().length > 0) {
+					this._applet_context_menu.toggle();
+				}
+			}
+		}
+		return true;
+  }, // End of _onButtonPressEvent
 
   //++ Handler for when the applet is clicked.
   on_applet_clicked:function (event) {